Ultimos detalles de Css, se modulo el codigo
