+++ src/index.html
@@ -57,7 +57,7 @@
     <!--RESULTADOS BUSCADOR-->
 
     <div id="modalPoke" class="modalBuscador">
-      <div id="flexPoke" class="flexBuscador">
+      <div class="fondo">
         <div class="contenido-modal-buscador">
           <div id="circle">
             <span class="closeBuscador" id="closePoke">&times;</span>
